refactor Invoices to component
